requirements.txt
yt-dlp~=2025.7.21
python-telegram-bot~=20.7
python-dotenv~=1.1.1